Add files via upload
In this notebook, we will explore the fundamentals of using a lakehouse and Apache Spark for data storage and analysis. We will showcase the process of accessing data from the lakehouse, analyzing it with Spark DataFrames, executing elementary data transformations, and preparing the data for machine learning models. Upon completion of this guide, you will possess a solid comprehension of utilizing a lakehouse for data storage and analysis, as well as be adept at executing routine data engineering tasks using Spark. Let's begin!